split model_object_def into model_factory and basic_module_def
